Refactor shop detection helpers
